requirements.txt
pytest
pillow
